ZHA: Create Route API Auth and Change Design of auth register
