+++ with-docker-compose-app/next-app/logs/combined.log
@@ -96,3 +96,11 @@
 {"level":"error","message":"Une erreur fictive est survenue - Niveau Error","timestamp":"2024-10-21T14:02:46.173Z"}
 {"level":"info","message":"Page d'accueil rendue avec succès - Niveau Info","timestamp":"2024-10-21T15:07:14.847Z"}
 {"level":"error","message":"Une erreur fictive est survenue - Niveau Error","timestamp":"2024-10-21T15:07:14.849Z"}
+{"level":"info","message":"Page d'accueil rendue avec succès - Niveau Info","timestamp":"2024-10-21T15:38:13.338Z"}
+{"level":"error","message":"Une erreur fictive est survenue - Niveau Error","timestamp":"2024-10-21T15:38:13.340Z"}
+{"level":"info","message":"Page d'accueil rendue avec succès - Niveau Info","timestamp":"2024-10-21T15:38:27.827Z"}
+{"level":"error","message":"Une erreur fictive est survenue - Niveau Error","timestamp":"2024-10-21T15:38:27.827Z"}
+{"level":"info","message":"Page d'accueil rendue avec succès - Niveau Info","timestamp":"2024-10-21T15:49:51.619Z"}
+{"level":"error","message":"Une erreur fictive est survenue - Niveau Error","timestamp":"2024-10-21T15:49:51.621Z"}
+{"level":"info","message":"Page d'accueil rendue avec succès - Niveau Info","timestamp":"2024-10-21T16:03:44.148Z"}
+{"level":"error","message":"Une erreur fictive est survenue - Niveau Error","timestamp":"2024-10-21T16:03:44.150Z"}
