+++ with-docker-compose-app/next-app/logs/error.log
@@ -16,3 +16,7 @@
 {"level":"error","message":"Une erreur fictive est survenue - Niveau Error","timestamp":"2024-10-21T09:18:36.318Z"}
 {"level":"error","message":"Une erreur fictive est survenue - Niveau Error","timestamp":"2024-10-21T14:02:46.173Z"}
 {"level":"error","message":"Une erreur fictive est survenue - Niveau Error","timestamp":"2024-10-21T15:07:14.849Z"}
+{"level":"error","message":"Une erreur fictive est survenue - Niveau Error","timestamp":"2024-10-21T15:38:13.340Z"}
+{"level":"error","message":"Une erreur fictive est survenue - Niveau Error","timestamp":"2024-10-21T15:38:27.827Z"}
+{"level":"error","message":"Une erreur fictive est survenue - Niveau Error","timestamp":"2024-10-21T15:49:51.621Z"}
+{"level":"error","message":"Une erreur fictive est survenue - Niveau Error","timestamp":"2024-10-21T16:03:44.150Z"}
