+++ with-docker-compose-app/next-app/src/components/Header/Logo.js
@@ -5,16 +5,6 @@ import Image from "next/image";
 
 function Logo() {
   return (
-    // <Link href="/">
-    //       <Image
-    //         src="/images/logo.svg"
-    //         alt="NSK Site"
-    //         width={100}
-    //         height={40}
-    //         className={styles.logo}
-    //       />
-    //
-
     <Link href="/">
       <h2 className="fw-bold text-white">NSK Site</h2>
     </Link>
