+++ with-docker-compose-app/next-app/src/pages/auth/index.js
@@ -1,0 +1,17 @@
+import AuthForm from '../../components/Authentification/AuthForm';
+
+function AuthPage() {
+  return (
+    <div className="container" id="auth">
+        <h1 className="text-center text-dark pt-5">
+            Inscription
+        </h1> 
+        <a href="/login" className="text-center text-dark pt-2 d-block">
+           <u>Si vous avez déjà un compte, veuillez vous connecter</u> 
+        </a>
+        <AuthForm />
+    </div>
+  );
+}
+
+export default AuthPage;
