--- with-docker-compose-app/next-app/src/pages/login/index.js
@@ -1,9 +1,0 @@
-import Login from "../../components/Authentification/Login";
-
-export default  function LoginPage() {
-  return (
-    <>
-      <Login />
-    </>
-  );
-};
