--- with-docker-compose-app/next-app/src/pages/register/index.js
@@ -1,9 +1,0 @@
-import Register from "../../components/Authentification/Register";
-
-export default  function RegisterPage() {
-  return (
-    <>
-      <Register />
-    </>
-  );
-};
