+++ with-docker-compose-app/next-app/src/styles/_variables.css
@@ -15,6 +15,7 @@
     --grayish-blue: hsl(240, 2%, 79%);
     --very-dark-grayish-blue: hsl(207, 13%, 34%);
     --very-dark-black-blue: hsl(240, 10%, 16%);
+    --very-grey-blue: rgb(239 246 255 / 1);
     /* Gradient Colors */
     --gradient-mobile-nav-start: hsl(13, 100%, 72%);
     --gradient-mobile-nav-end: hsl(353, 100%, 62%);
