+++ with-docker-compose-app/next-app/src/styles/globals.css
@@ -20,3 +20,5 @@ a {
   color: inherit;
   text-decoration: none;
 }
+
+
